+++ Addon/ElvUI_mMediaTag/core/changelog/3.58.lua
@@ -1,0 +1,7 @@
+mMT.Changelog[358]  = {
+	DATE = "11.09.2024",
+	NEW = {
+		"[Portraits]: Add a function to add your custom Class Icons to the portraits via code.",
+		"[Portraits]: Add e Set of new Class Icons.",
+	},
+}