+++ Addon/ElvUI_mMediaTag/core/options/portraits.lua
@@ -58,15 +58,18 @@ local frameStrata = {
 }
 
 function BuildIconStylesTable()
-	for id, name in pairs(mMT.ElvUI_JiberishIcons.styles) do
-		ClassIconStyle[id] = name
+	-- for id, name in pairs(mMT.ElvUI_JiberishIcons.styles) do
+	-- 	ClassIconStyle[id] = name
+	-- end
+	for iconStyle, value in pairs(mMT.classIcons) do
+		ClassIconStyle[iconStyle] = value.name
 	end
 end
 
 local function configTable()
 	local sizeString = ":16:16:0:0:64:64:4:60:4:60"
 
-	if mMT.ElvUI_JiberishIcons.loaded then BuildIconStylesTable() end
+	BuildIconStylesTable()
 
 	E.Options.args.mMT.args.unitframes.args.portraits.args = {
 		toggle_enable = {
@@ -123,6 +126,7 @@ local function configTable()
 									end,
 									set = function(info, value)
 										E.db.mMT.portraits.general.classiconstyle = value
+										mMT.Modules.Portraits:Initialize()
 										E:StaticPopup_Show("CONFIG_RL")
 									end,
 									values = ClassIconStyle,